fix: all page path
