+++ components/Layouts.tsx
@@ -7,7 +7,7 @@ const Layout = ({children} : {children: JSX.Element}) => {
     <div className={styles.wrapper}>
       <Header />
       <div className={styles.cv_tag_wrapper}>
-        <CvTag link="/somei_juku_sample/about" content="資料請求・体験授業のご予約受付中" />
+        <CvTag link="/about" content="資料請求・体験授業のご予約受付中" />
       </div>
       { children }
       <Footer />
