+++ pages/assets/curriculum_card_2.png
