+++ pages/assets/curriculum_title_image.png
