+++ pages/home.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../styles/variables';
 
 .section_top {
   margin-top: 62px;
