--- pages/somei_juku_sample/plan.tsx
@@ -1,11 +1,0 @@
-import Layout from "../../components/Layouts"
-
-const Page = () => {
-  return (
-    <div>Plan</div>
-  )
-}
-
-Page.getLayout = (page: JSX.Element) => <Layout>{page}</Layout>
-
-export default Page
